Finished connector/helper refactoring and restructured the directory to reflect the new abstractions. This should be a lot better than it was.
